requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.1.0
bleach==4.1.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.1.0
crayons==0.4.0
cryptography==35.0.0
cycler==0.10.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
EbookLib==0.17.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.2
filelock==3.7.0
gTTS==2.2.3
h11==0.12.0
idna==3.3
ipykernel==6.8.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.2
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
mplfinance==0.12.7a17
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.2
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pandas==1.2.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2022.5.2
platformdirs==2.4.1
playsound==1.3.0
plotly==5.3.1
praw==7.4.0
prawcore==2.3.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.20
Pygments==2.11.2
pyobjc==8.2
pyobjc-core==8.2
pyobjc-framework-Accessibility==8.2
pyobjc-framework-Accounts==8.2
pyobjc-framework-AddressBook==8.2
pyobjc-framework-AdServices==8.2
pyobjc-framework-AdSupport==8.2
pyobjc-framework-AppleScriptKit==8.2
pyobjc-framework-AppleScriptObjC==8.2
pyobjc-framework-ApplicationServices==8.2
pyobjc-framework-AppTrackingTransparency==8.2
pyobjc-framework-AudioVideoBridging==8.2
pyobjc-framework-AuthenticationServices==8.2
pyobjc-framework-AutomaticAssessmentConfiguration==8.2
pyobjc-framework-Automator==8.2
pyobjc-framework-AVFoundation==8.2
pyobjc-framework-AVKit==8.2
pyobjc-framework-BusinessChat==8.2
pyobjc-framework-CalendarStore==8.2
pyobjc-framework-CallKit==8.2
pyobjc-framework-CFNetwork==8.2
pyobjc-framework-ClassKit==8.2
pyobjc-framework-CloudKit==8.2
pyobjc-framework-Cocoa==8.2
pyobjc-framework-Collaboration==8.2
pyobjc-framework-ColorSync==8.2
pyobjc-framework-Contacts==8.2
pyobjc-framework-ContactsUI==8.2
pyobjc-framework-CoreAudio==8.2
pyobjc-framework-CoreAudioKit==8.2
pyobjc-framework-CoreBluetooth==8.2
pyobjc-framework-CoreData==8.2
pyobjc-framework-CoreHaptics==8.2
pyobjc-framework-CoreLocation==8.2
pyobjc-framework-CoreMedia==8.2
pyobjc-framework-CoreMediaIO==8.2
pyobjc-framework-CoreMIDI==8.2
pyobjc-framework-CoreML==8.2
pyobjc-framework-CoreMotion==8.2
pyobjc-framework-CoreServices==8.2
pyobjc-framework-CoreSpotlight==8.2
pyobjc-framework-CoreText==8.2
pyobjc-framework-CoreWLAN==8.2
pyobjc-framework-CryptoTokenKit==8.2
pyobjc-framework-DataDetection==8.2
pyobjc-framework-DeviceCheck==8.2
pyobjc-framework-DictionaryServices==8.2
pyobjc-framework-DiscRecording==8.2
pyobjc-framework-DiscRecordingUI==8.2
pyobjc-framework-DiskArbitration==8.2
pyobjc-framework-DVDPlayback==8.2
pyobjc-framework-EventKit==8.2
pyobjc-framework-ExceptionHandling==8.2
pyobjc-framework-ExecutionPolicy==8.2
pyobjc-framework-ExternalAccessory==8.2
pyobjc-framework-FileProvider==8.2
pyobjc-framework-FileProviderUI==8.2
pyobjc-framework-FinderSync==8.2
pyobjc-framework-FSEvents==8.2
pyobjc-framework-GameCenter==8.2
pyobjc-framework-GameController==8.2
pyobjc-framework-GameKit==8.2
pyobjc-framework-GameplayKit==8.2
pyobjc-framework-ImageCaptureCore==8.2
pyobjc-framework-IMServicePlugIn==8.2
pyobjc-framework-InputMethodKit==8.2
pyobjc-framework-InstallerPlugins==8.2
pyobjc-framework-InstantMessage==8.2
pyobjc-framework-Intents==8.2
pyobjc-framework-IntentsUI==8.2
pyobjc-framework-IOSurface==8.2
pyobjc-framework-iTunesLibrary==8.2
pyobjc-framework-KernelManagement==8.2
pyobjc-framework-LatentSemanticMapping==8.2
pyobjc-framework-LaunchServices==8.2
pyobjc-framework-libdispatch==8.2
pyobjc-framework-LinkPresentation==8.2
pyobjc-framework-LocalAuthentication==8.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.2
pyobjc-framework-MailKit==8.2
pyobjc-framework-MapKit==8.2
pyobjc-framework-MediaAccessibility==8.2
pyobjc-framework-MediaLibrary==8.2
pyobjc-framework-MediaPlayer==8.2
pyobjc-framework-MediaToolbox==8.2
pyobjc-framework-Metal==8.2
pyobjc-framework-MetalKit==8.2
pyobjc-framework-MetalPerformanceShaders==8.2
pyobjc-framework-MetalPerformanceShadersGraph==8.2
pyobjc-framework-MetricKit==8.2
pyobjc-framework-MLCompute==8.2
pyobjc-framework-ModelIO==8.2
pyobjc-framework-MultipeerConnectivity==8.2
pyobjc-framework-NaturalLanguage==8.2
pyobjc-framework-NetFS==8.2
pyobjc-framework-Network==8.2
pyobjc-framework-NetworkExtension==8.2
pyobjc-framework-NotificationCenter==8.2
pyobjc-framework-OpenDirectory==8.2
pyobjc-framework-OSAKit==8.2
pyobjc-framework-OSLog==8.2
pyobjc-framework-PassKit==8.2
pyobjc-framework-PencilKit==8.2
pyobjc-framework-Photos==8.2
pyobjc-framework-PhotosUI==8.2
pyobjc-framework-PreferencePanes==8.2
pyobjc-framework-PushKit==8.2
pyobjc-framework-Quartz==8.2
pyobjc-framework-QuickLookThumbnailing==8.2
pyobjc-framework-ReplayKit==8.2
pyobjc-framework-SafariServices==8.2
pyobjc-framework-SceneKit==8.2
pyobjc-framework-ScreenSaver==8.2
pyobjc-framework-ScreenTime==8.2
pyobjc-framework-ScriptingBridge==8.2
pyobjc-framework-SearchKit==8.2
pyobjc-framework-Security==8.2
pyobjc-framework-SecurityFoundation==8.2
pyobjc-framework-SecurityInterface==8.2
pyobjc-framework-ServiceManagement==8.2
pyobjc-framework-ShazamKit==8.2
pyobjc-framework-Social==8.2
pyobjc-framework-SoundAnalysis==8.2
pyobjc-framework-Speech==8.2
pyobjc-framework-SpriteKit==8.2
pyobjc-framework-StoreKit==8.2
pyobjc-framework-SyncServices==8.2
pyobjc-framework-SystemConfiguration==8.2
pyobjc-framework-SystemExtensions==8.2
pyobjc-framework-UniformTypeIdentifiers==8.2
pyobjc-framework-UserNotifications==8.2
pyobjc-framework-UserNotificationsUI==8.2
pyobjc-framework-VideoSubscriberAccount==8.2
pyobjc-framework-VideoToolbox==8.2
pyobjc-framework-Virtualization==8.2
pyobjc-framework-Vision==8.2
pyobjc-framework-WebKit==8.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.27.12
pypub==1.5
pyrsistent==0.18.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2021.3
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.26.0
selenium==4.0.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
stack-data==0.1.4
tenacity==8.0.1
terminado==0.13.1
testpath==0.5.0
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
update-checker==0.18.0
urllib3==1.26.7
virtualenv==20.14.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
webdriver-manager==3.5.2
webencodings==0.5.1
websocket-client==1.2.1
widgetsnbextension==3.5.2
wsproto==1.0.0
